+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/deprecated/data/JsonDsl.kt
@@ -1,11 +1,12 @@
-package com.windea.breezeframework.data.dsl.deprecated.data
+package com.windea.breezeframework.dsl.deprecated.data
 
 import com.windea.breezeframework.core.annotations.marks.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.data.dsl.deprecated.*
-import com.windea.breezeframework.data.dsl.deprecated.data.JsonDslConfig.defaultName
-import com.windea.breezeframework.data.dsl.deprecated.data.JsonDslConfig.indent
-import com.windea.breezeframework.data.dsl.deprecated.data.JsonDslConfig.quote
+import com.windea.breezeframework.dsl.deprecated.*
+import com.windea.breezeframework.dsl.deprecated.data.JsonDslConfig.defaultName
+import com.windea.breezeframework.dsl.deprecated.data.JsonDslConfig.indent
+import com.windea.breezeframework.dsl.deprecated.data.JsonDslConfig.quote
 import java.lang.annotation.*
 
 /**Json Dsl。*/