+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/deprecated/data/YamlDsl.kt
@@ -1,4 +1,4 @@
-package com.windea.breezeframework.data.dsl.deprecated.data
+package com.windea.breezeframework.dsl.deprecated.data
 
 //Yaml是Json的超集，但是需要重载toString()方法
 //DELAY