+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/mermaid/MermaidFlowDsl.kt
@@ -1,67 +1,36 @@
-@file:Reference("[Mermaid](https://mermaidjs.github.io)")
+@file:Reference("[Mermaid Flow Chart](https://mermaidjs.github.io/#/flowchart)")
 @file:Suppress("CanBePrimaryConstructorProperty", "NOTHING_TO_INLINE")
 
-package com.windea.breezeframework.data.dsl.text
+package com.windea.breezeframework.dsl.mermaid
 
 import com.windea.breezeframework.core.annotations.marks.*
 import com.windea.breezeframework.core.extensions.*
-import com.windea.breezeframework.data.dsl.*
-import com.windea.breezeframework.data.dsl.text.MermaidFlowConfig.indent
-import com.windea.breezeframework.data.dsl.text.MermaidFlowConfig.quote
+import com.windea.breezeframework.dsl.*
+import com.windea.breezeframework.dsl.mermaid.MermaidFlowConfig.indent
+import com.windea.breezeframework.dsl.mermaid.MermaidFlowConfig.quote
 
-
-//REGION Dsl marker annotations & Dsl element interfaces
+//REGION Dsl annotations
 
 @DslMarker
 annotation class MermaidFlowDsl
 
+//REGION Dsl elements & build functions
+
+/**构建Mermaid流程图。*/
+@MermaidFlowDsl
+fun mermaidFlow(direction: MermaidFlowDirection, builder: MermaidFlow.() -> Unit) =
+	MermaidFlow(direction).also { it.builder() }
+
+/**Mermaid流程图Dsl的元素。*/
 @MermaidFlowDsl
 interface MermaidFlowDslElement
 
+/**抽象的Mermaid流程图。*/
 @MermaidFlowDsl
-abstract class MermaidFlowGraph : CanIndentContent {
+abstract class AbstractMermaidFlow : MermaidFlowDslElement, CanIndentContent {
 	val nodes: MutableSet<MermaidFlowNode> = mutableSetOf()
 	val links: MutableSet<MermaidFlowLink> = mutableSetOf()
 	val subGraphs: MutableList<MermaidFlowSubGraph> = mutableListOf()
-	
-	override var indentContent: Boolean = true
-	
-	@MermaidFlowDsl
-	inline fun node(id: String, text: String? = null, shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect) =
-		MermaidFlowNode(id, text, shape).also { nodes += it }
-	
-	@MermaidFlowDsl
-	inline fun link(sourceNodeId: String, targetNodeId: String, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) =
-		MermaidFlowLink(sourceNodeId, targetNodeId, linkShape, linkText).also { links += it }
-	
-	@MermaidFlowDsl
-	inline fun link(sourceNode: MermaidFlowNode, targetNodeId: String, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) =
-		MermaidFlowLink(sourceNode.id, targetNodeId, linkShape, linkText).also { links += it }
-	
-	@MermaidFlowDsl
-	inline fun link(sourceNodeId: String, targetNode: MermaidFlowNode, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) =
-		MermaidFlowLink(sourceNodeId, targetNode.id, linkShape, linkText).also { links += it }
-	
-	@MermaidFlowDsl
-	inline fun link(sourceNode: MermaidFlowNode, targetNode: MermaidFlowNode, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) =
-		MermaidFlowLink(sourceNode.id, targetNode.id, linkShape, linkText).also { links += it }
-	
-	@MermaidFlowDsl
-	inline fun subGraph(name: String, builder: MermaidFlowSubGraph.() -> Unit) =
-		MermaidFlowSubGraph(name).also { it.builder() }.also { subGraphs += it }
-}
-
-//REGION Dsl elements & build functions
-
-@MermaidFlowDsl
-fun mermaidGraph(direction: MermaidFlowDirection, builder: MermaidFlow.() -> Unit) =
-	MermaidFlow(direction).also { it.builder() }
-
-@MermaidFlowDsl
-class MermaidFlow @PublishedApi internal constructor(
-	direction: MermaidFlowDirection
-) : MermaidFlowGraph(), DslBuilder {
-	val direction: MermaidFlowDirection = direction
 	val styles: MutableList<MermaidFlowNodeStyle> = mutableListOf()
 	val linkStyles: MutableList<MermaidFlowLinkStyle> = mutableListOf()
 	val classDefs: MutableSet<MermaidFlowClassDef> = mutableSetOf()
@@ -70,7 +39,7 @@ class MermaidFlow @PublishedApi internal constructor(
 	override var indentContent: Boolean = true
 	
 	override fun toString(): String {
-		val contentSnippet = arrayOf(
+		return arrayOf(
 			nodes.joinToString("\n"),
 			links.joinToString("\n"),
 			subGraphs.joinToString("\n"),
@@ -79,47 +48,65 @@ class MermaidFlow @PublishedApi internal constructor(
 			classDefs.joinToString("\n"),
 			classRefs.joinToString("\n")
 		).filterNotEmpty().joinToString("\n\n")
-		val indentedSnippet = if(indentContent) contentSnippet.prependIndent(indent) else contentSnippet
-		return "graph ${direction.text}\n$indentedSnippet"
 	}
 	
+	
+	@MermaidFlowDsl
+	inline fun node(id: String, text: String? = null, shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect) =
+		MermaidFlowNode(id, text, shape).also { nodes += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNodeId: String, targetNodeId: String, linkText: String? = null, linkArrow: MermaidFlowLinkArrow = MermaidFlowLinkArrow.Arrow) =
+		MermaidFlowLink(sourceNodeId, targetNodeId, linkText, linkArrow).also { links += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNode: MermaidFlowNode, targetNodeId: String, linkText: String? = null, linkArrow: MermaidFlowLinkArrow = MermaidFlowLinkArrow.Arrow) =
+		MermaidFlowLink(sourceNode.id, targetNodeId, linkText, linkArrow).also { links += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNodeId: String, targetNode: MermaidFlowNode, linkText: String? = null, linkArrow: MermaidFlowLinkArrow = MermaidFlowLinkArrow.Arrow) =
+		MermaidFlowLink(sourceNodeId, targetNode.id, linkText, linkArrow).also { links += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNode: MermaidFlowNode, targetNode: MermaidFlowNode, linkText: String? = null, linkArrow: MermaidFlowLinkArrow = MermaidFlowLinkArrow.Arrow) =
+		MermaidFlowLink(sourceNode.id, targetNode.id, linkText, linkArrow).also { links += it }
+	
+	@MermaidFlowDsl
+	inline fun subGraph(name: String, builder: MermaidFlowSubGraph.() -> Unit) =
+		MermaidFlowSubGraph(name).also { it.builder() }.also { subGraphs += it }
 	
 	@MermaidFlowDsl
 	inline fun style(nodeId: String, vararg styles: Pair<String, String>) =
-		MermaidFlowNodeStyle(nodeId, *styles).also { this@MermaidFlow.styles += it }
+		MermaidFlowNodeStyle(nodeId, *styles).also { this@AbstractMermaidFlow.styles += it }
 	
 	@MermaidFlowDsl
 	inline fun linkStyle(linkOrder: Int, vararg styles: Pair<String, String>) =
-		MermaidFlowLinkStyle(linkOrder, *styles).also { this@MermaidFlow.linkStyles += it }
+		MermaidFlowLinkStyle(linkOrder, *styles).also { linkStyles += it }
 	
 	@MermaidFlowDsl
 	inline fun classDef(className: String, vararg styles: Pair<String, String>) =
-		MermaidFlowClassDef(className, *styles).also { this@MermaidFlow.classDefs += it }
+		MermaidFlowClassDef(className, *styles).also { classDefs += it }
 	
 	@MermaidFlowDsl
 	inline fun classRef(vararg nodeIds: String, className: String) =
-		MermaidFlowClassRef(*nodeIds, className = className).also { this@MermaidFlow.classRefs += it }
+		MermaidFlowClassRef(*nodeIds, className = className).also { classRefs += it }
 }
 
+/**Mermaid流程图。*/
 @MermaidFlowDsl
-class MermaidFlowSubGraph @PublishedApi internal constructor(
-	name: String
-) : MermaidFlowGraph() {
-	val name: String = name //NOTE do not ensure argument is valid
-	
-	override var indentContent: Boolean = true
+class MermaidFlow @PublishedApi internal constructor(
+	direction: MermaidFlowDirection
+) : AbstractMermaidFlow(), Dsl {
+	val direction: MermaidFlowDirection = direction
 	
 	override fun toString(): String {
-		val contentSnippet = arrayOf(
-			nodes.joinToString("\n"),
-			links.joinToString("\n"),
-			subGraphs.joinToString("\n")
-		).filterNotEmpty().joinToString("\n\n")
+		val contentSnippet = super.toString()
 		val indentedSnippet = if(indentContent) contentSnippet.prependIndent(indent) else contentSnippet
-		return "subgraph $name\n$indentedSnippet\nend"
+		return "graph ${direction.text}\n$indentedSnippet"
 	}
 }
 
+/**Mermaid流程图节点。*/
 @MermaidFlowDsl
 class MermaidFlowNode @PublishedApi internal constructor(
 	id: String,
@@ -127,7 +114,7 @@ class MermaidFlowNode @PublishedApi internal constructor(
 	shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect
 ) : MermaidFlowDslElement {
 	val id: String = id //NOTE do not ensure argument is valid
-	val text: String? = text //NOTE do not ensure argument is valid
+	val text: String? = text?.replaceWithHtmlWrap() //NOTE do not ensure argument is valid
 	val shape: MermaidFlowNodeShape = shape
 	
 	//TODO omit text, no double quote surround text if not necessary
@@ -137,26 +124,42 @@ class MermaidFlowNode @PublishedApi internal constructor(
 	}
 }
 
+/**Mermaid流程图连接。*/
 @MermaidFlowDsl
 class MermaidFlowLink @PublishedApi internal constructor(
 	sourceNodeId: String,
 	targetNodeId: String,
-	linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow,
-	linkText: String? = null
+	linkText: String? = null,
+	linkArrow: MermaidFlowLinkArrow = MermaidFlowLinkArrow.Arrow
 ) : MermaidFlowDslElement {
 	val sourceNodeId: String = sourceNodeId //NOTE do not ensure argument is valid
 	val targetNodeId: String = targetNodeId //NOTE do not ensure argument is valid
-	val linkShape: MermaidFlowLinkShape = linkShape
-	val linkText: String? = linkText //NOTE do not ensure argument is valid
+	val linkText: String? = linkText?.replaceWithHtmlWrap() //NOTE do not ensure argument is valid
+	val linkArrow: MermaidFlowLinkArrow = linkArrow
 	
 	//TODO escape if necessary
 	override fun toString(): String {
-		val linkSnippet = linkShape.delimiter
+		val linkSnippet = linkArrow.text
 		val linkTextSnippet = linkText?.let { "|$linkText|" } ?: ""
 		return "$sourceNodeId $linkSnippet$linkTextSnippet $targetNodeId"
 	}
 }
 
+/**Mermaid流程图的子图。*/
+@MermaidFlowDsl
+class MermaidFlowSubGraph @PublishedApi internal constructor(
+	name: String
+) : AbstractMermaidFlow() {
+	val name: String = name //NOTE do not ensure argument is valid
+	
+	override fun toString(): String {
+		val contentSnippet = super.toString()
+		val indentedSnippet = if(indentContent) contentSnippet.prependIndent(indent) else contentSnippet
+		return "subgraph $name\n$indentedSnippet\nend"
+	}
+}
+
+/**Mermaid流程图节点风格。*/
 @MermaidFlowDsl
 class MermaidFlowNodeStyle @PublishedApi internal constructor(
 	nodeId: String,
@@ -171,6 +174,7 @@ class MermaidFlowNodeStyle @PublishedApi internal constructor(
 	}
 }
 
+/**Mermaid流程图连接风格。*/
 @MermaidFlowDsl
 class MermaidFlowLinkStyle @PublishedApi internal constructor(
 	linkOrder: Int,
@@ -185,6 +189,7 @@ class MermaidFlowLinkStyle @PublishedApi internal constructor(
 	}
 }
 
+/**Mermaid流程图类定义。*/
 @MermaidFlowDsl
 class MermaidFlowClassDef @PublishedApi internal constructor(
 	className: String,
@@ -199,6 +204,7 @@ class MermaidFlowClassDef @PublishedApi internal constructor(
 	}
 }
 
+/**Mermaid流程图类引用。*/
 @MermaidFlowDsl
 class MermaidFlowClassRef @PublishedApi internal constructor(
 	vararg nodeIds: String,
@@ -215,12 +221,14 @@ class MermaidFlowClassRef @PublishedApi internal constructor(
 
 //REGION Enumerations and constants
 
+/**Mermaid流程图的方向。*/
 enum class MermaidFlowDirection(
 	val text: String
 ) {
 	TB("TB"), BT("BT"), LR("LR"), RL("RL")
 }
 
+/**Mermaid流程图节点的形状。*/
 enum class MermaidFlowNodeShape(
 	val prefix: String,
 	val suffix: String
@@ -237,17 +245,21 @@ enum class MermaidFlowNodeShape(
 	Rhombus("{", "}")
 }
 
-enum class MermaidFlowLinkShape(
-	val delimiter: String
+/**Mermaid流程图连接的箭头类型。*/
+enum class MermaidFlowLinkArrow(
+	val text: String
 ) {
 	Arrow("-->"),
-	Dotted("-.->"),
-	Thick("==>"),
-	Normal("---")
+	DottedArrow("-.->"),
+	BoldArrow("==>"),
+	Line("---"),
+	DottedLine("-.-"),
+	BoldLine("===")
 }
 
 //REGION Config object
 
+/**Mermaid流程图的配置。*/
 object MermaidFlowConfig : DslConfig {
 	/**缩进长度。*/
 	var indentSize = 2